Error check for drive strength on input-only pins
