+++ FluidNC/esp32/GPIOCapabilities.cpp
@@ -1,0 +1,71 @@
+#include "src/Pins/GPIOPinDetail.h"
+
+namespace Pins {
+    PinCapabilities GPIOPinDetail::GetDefaultCapabilities(pinnum_t index) {
+        // See https://randomnerdtutorials.com/esp32-pinout-reference-gpios/ for an overview:
+        switch (index) {
+            case 0:  // Outputs PWM signal at boot
+                return PinCapabilities::Native | PinCapabilities::Input | PinCapabilities::Output | PinCapabilities::PullUp |
+                       PinCapabilities::PullDown | PinCapabilities::ADC | PinCapabilities::PWM | PinCapabilities::ISR |
+                       PinCapabilities::UART;
+
+            case 1:  // TX pin of Serial0. Note that Serial0 also runs through the Pins framework!
+                return PinCapabilities::Native | PinCapabilities::Output | PinCapabilities::Input | PinCapabilities::UART;
+
+            case 3:  // RX pin of Serial0. Note that Serial0 also runs through the Pins framework!
+                return PinCapabilities::Native | PinCapabilities::Output | PinCapabilities::Input | PinCapabilities::ISR |
+                       PinCapabilities::UART;
+
+            case 5:
+            case 9:
+            case 10:
+            case 16:
+            case 17:
+            case 18:
+            case 19:
+            case 21:
+            case 22:
+            case 23:
+            case 29:
+                return PinCapabilities::Native | PinCapabilities::Input | PinCapabilities::Output | PinCapabilities::PullUp |
+                       PinCapabilities::PullDown | PinCapabilities::PWM | PinCapabilities::ISR | PinCapabilities::UART;
+
+            case 2:  // Normal pins
+            case 4:
+            case 12:  // Boot fail if pulled high
+            case 13:
+            case 14:  // Outputs PWM signal at boot
+            case 15:  // Outputs PWM signal at boot
+            case 27:
+            case 32:
+            case 33:
+                return PinCapabilities::Native | PinCapabilities::Input | PinCapabilities::Output | PinCapabilities::PullUp |
+                       PinCapabilities::PullDown | PinCapabilities::ADC | PinCapabilities::PWM | PinCapabilities::ISR |
+                       PinCapabilities::UART;
+
+            case 25:
+            case 26:
+                return PinCapabilities::Native | PinCapabilities::Input | PinCapabilities::Output | PinCapabilities::PullUp |
+                       PinCapabilities::PullDown | PinCapabilities::ADC | PinCapabilities::DAC | PinCapabilities::PWM |
+                       PinCapabilities::ISR | PinCapabilities::UART;
+
+            case 6:  // SPI flash integrated
+            case 7:
+            case 8:
+            case 11:
+                return PinCapabilities::Reserved;
+
+            case 34:  // Input only pins
+            case 35:
+            case 36:
+            case 37:
+            case 38:
+            case 39:
+                return PinCapabilities::Native | PinCapabilities::Input | PinCapabilities::ADC | PinCapabilities::ISR | PinCapabilities::UART;
+                break;
+
+            default:  // Not mapped to actual GPIO pins
+                return PinCapabilities::None;
+        }
+    }
+}
