+++ FluidNC/src/Pins/PinAttributes.cpp
@@ -2,7 +2,6 @@
 // Use of this source code is governed by a GPLv3 license that can be found in the LICENSE file.
 
 #include "PinAttributes.h"
-#include "PinCapabilities.h"
 
 namespace Pins {
     PinAttributes PinAttributes::Undefined(0);
