+++ FluidNC/src/Pins/PinAttributes.h
@@ -4,6 +4,7 @@
 #pragma once
 
 #include <cstdint>
+#include "Driver/fluidnc_gpio.h"
 
 #include "PinCapabilities.h"
